services/game-service/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-chess==1.999
requests==2.31.0